LayoutTests/plugins/npp-set-window-called-during-destruction-expected.txt

CONSOLE MESSAGE: PLUGIN: Success: NPP_SetWindow was called during plugin destruction
This tests whether a NPP_SetWindow is called during plugin destruction.
